+++ index.js
@@ -1,10 +1,10 @@
 /**
  * Created by aurum on 2018/3/14.
  */
-const GatewayPay = require('./gateway_pay');
-const AccountPay = require('./account_pay');
-const Shouyinbao = require('./shouyinbao');
-const Wanjiantong = require('./wanjiantong');
+const GatewayPay = require('./libs/gateway_pay');
+const AccountPay = require('./libs/account_pay');
+const Shouyinbao = require('./libs/shouyinbao');
+const Wanjiantong = require('./libs/wanjiantong');
 
 module.exports = {
     GatewayPay,
