+++ libs/shouyinbao.js
